requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
torch
torchvision
opencv-python
pillow
numpy
